Add components that handle user login/register and bug fix
JS/JSX - RegisterOrLogin will be wired up to allow users to make http requests to the backend. CheckForUser will be wired to check for registereduser on backend and route the frontend user accordingly. Keys on icon components changed to unique strings to match name of icon. Math.random() assigned to key was creating component re-render issues on drawer open/close because of dynamic key